Update zero icon asset and project structure
