+++ package.json
@@ -71,7 +71,7 @@
 	},
 	"dependencies": {
 		"@zero-tech/zns-sdk": "0.8.9",
-		"@zero-tech/zui": "0.7.0",
+		"@zero-tech/zui": "0.11.4",
 		"classnames": "^2.3.1",
 		"react-query": "^3.39.1"
 	},
