+++ src/features/buy-domain/BuyDomainModal/BuyDomainModal.module.scss
@@ -34,22 +34,6 @@
 		.WizardContent {
 			padding: 0 1.625rem;
 
-			.DefaultDomainMedia {
-				position: relative;
-				width: 100%;
-				margin: 2.5rem 0;
-				padding-top: 56.25%; /* 16:9 Aspect Ratio */
-				border-radius: 0.5rem;
-				overflow: hidden;
-
-				video {
-					position: absolute;
-					top: 0;
-					height: 100%;
-					object-fit: cover;
-				}
-			}
-
 			.WizardContentSection {
 				:global(.zui-loading-indicator) {
 					display: flex;
@@ -57,6 +41,10 @@
 					align-items: center;
 					gap: 2rem;
 					text-align: center;
+				}
+
+				:global(.zui-loading-indicator-text) {
+					display: inline-flex;
 				}
 
 				.Confirmation {
