+++ src/features/buy-domain/BuyDomainModal/Steps/index.ts
@@ -1,4 +1,3 @@
-export * from './DefaultDomainMedia';
 export * from './Discover';
 export * from './Approve';
 export * from './Mint';
