--- src/features/buy-domain/Icons/IconZero.tsx
@@ -1,57 +1,0 @@
-import type { FC } from 'react';
-
-export const IconZero: FC = () => (
-	<svg
-		width="36"
-		height="36"
-		viewBox="0 0 36 36"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<g filter="url(#filter0_d_8233_90083)">
-			<path
-				d="M6.3959 6.39595C0.156648 12.6353 0.608863 22.8936 7.43521 28.5651C2.56054 22.7247 2.83596 13.9205 8.37818 8.37826C13.9204 2.83597 22.7245 2.55942 28.5648 7.43527C22.8934 0.608848 12.6352 0.156624 6.3959 6.39595Z"
-				fill="white"
-			/>
-			<path
-				d="M29.6041 29.6044C35.8434 23.3651 35.3911 13.1068 28.5648 7.43527C33.4395 13.2756 33.164 22.0798 27.6218 27.6221C22.0796 33.1644 13.2755 33.441 7.43521 28.5651C13.1078 35.3904 23.366 35.8438 29.6041 29.6044Z"
-				fill="white"
-			/>
-		</g>
-		<defs>
-			<filter
-				id="filter0_d_8233_90083"
-				x="0"
-				y="0"
-				width="36"
-				height="36"
-				filterUnits="userSpaceOnUse"
-			>
-				<feFlood result="BackgroundImageFix" />
-				<feColorMatrix
-					in="SourceAlpha"
-					type="matrix"
-					values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-					result="hardAlpha"
-				/>
-				<feOffset />
-				<feGaussianBlur stdDeviation="1" />
-				<feColorMatrix
-					type="matrix"
-					values="0 0 0 0 1 0 0 0 0 1 0 0 0 0 1 0 0 0 0.5 0"
-				/>
-				<feBlend
-					mode="normal"
-					in2="BackgroundImageFix"
-					result="effect1_dropShadow_8233_90083"
-				/>
-				<feBlend
-					mode="normal"
-					in="SourceGraphic"
-					in2="effect1_dropShadow_8233_90083"
-					result="shape"
-				/>
-			</filter>
-		</defs>
-	</svg>
-);
