+++ src/features/buy-domain/index.ts
@@ -1,4 +1,3 @@
 export * from './BuyDomainButton';
 export * from './BuyDomainModal';
 export * from './DomainSearch';
-export * from './Icons';
