+++ src/features/ui/Icons/IconZero.tsx
@@ -1,0 +1,58 @@
+import type { FC } from 'react';
+
+export const IconZero: FC = () => (
+	<svg
+		width="45"
+		height="44"
+		viewBox="0 0 45 44"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+	>
+		<g filter="url(#filter0_d_3704_12868)">
+			<path
+				d="M7.99488 7.49494C0.195811 15.2941 0.761079 28.117 9.29401 35.2064C3.20068 27.9059 3.54494 16.9007 10.4727 9.97282C17.4005 3.04496 28.4056 2.69927 35.706 8.79409C28.6167 0.26106 15.7939 -0.304219 7.99488 7.49494Z"
+				fill="white"
+			/>
+			<path
+				d="M37.0051 36.5055C44.8042 28.7064 44.2389 15.8835 35.706 8.79409C41.7993 16.0946 41.4551 27.0998 34.5273 34.0277C27.5995 40.9555 16.5944 41.3012 9.29401 35.2064C16.3847 43.738 29.2075 44.3047 37.0051 36.5055Z"
+				fill="white"
+			/>
+		</g>
+		<defs>
+			<filter
+				id="filter0_d_3704_12868"
+				x="0.5"
+				y="0"
+				width="44"
+				height="44"
+				filterUnits="userSpaceOnUse"
+				color-interpolation-filters="sRGB"
+			>
+				<feFlood flood-opacity="0" result="BackgroundImageFix" />
+				<feColorMatrix
+					in="SourceAlpha"
+					type="matrix"
+					values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
+					result="hardAlpha"
+				/>
+				<feOffset />
+				<feGaussianBlur stdDeviation="1" />
+				<feColorMatrix
+					type="matrix"
+					values="0 0 0 0 1 0 0 0 0 1 0 0 0 0 1 0 0 0 0.5 0"
+				/>
+				<feBlend
+					mode="normal"
+					in2="BackgroundImageFix"
+					result="effect1_dropShadow_3704_12868"
+				/>
+				<feBlend
+					mode="normal"
+					in="SourceGraphic"
+					in2="effect1_dropShadow_3704_12868"
+					result="shape"
+				/>
+			</filter>
+		</defs>
+	</svg>
+);
