--- src/features/ui/QuestionButton/QuestionButton.module.scss
@@ -1,45 +1,0 @@
-@import '../../../../node_modules/@zero-tech//zui/styles/_animations.scss';
-@import '../../../../node_modules/@zero-tech/zui/styles/_borders.scss';
-@import '../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
-
-.QuestionButton {
-	position: relative;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	width: 16px;
-	height: 16px;
-	background: rgba($white, 0.25);
-	border-radius: 50%;
-	border: unset;
-	z-index: 1;
-
-	&::after {
-		content: ' ';
-		position: absolute;
-		top: 50%;
-		left: 50%;
-		height: 0px;
-		width: 0px;
-		border-radius: $radius-primary;
-		z-index: -2;
-		transition: box-shadow $time-medium ease-in-out;
-	}
-
-	&:hover::after {
-		box-shadow: 0 0 20px 13px rgba($white, 0.5);
-	}
-
-	img {
-		width: 6px;
-	}
-
-	&.Large {
-		width: 24px;
-		height: 24px;
-
-		img {
-			width: 8px;
-		}
-	}
-}
