--- src/features/ui/QuestionButton/index.ts
@@ -1,1 +1,0 @@
-export * from './QuestionButton';
